# waste/waste-client/src/main/resources/api-docs.yaml
openapi: 3.0.1
info:
  title: Yokudlela Storage API
  description: description =Yokudlela Waste Management API for Graphical User Interface
    .
  version: v1
servers:
- url: http://yokudlela:8081/waste
  description: local dev
paths:
  /add:
    post:
      tags:
      - waste-controller
      summary: Bevételezés
      operationId: addWaste
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Waste'
        required: true
      responses:
        "200":
          description: Sikeres hozzáadás
      security:
      - apikey:
        - waste
      - openid:
        - waste
      - oauth2:
        - waste
  /get:
    get:
      tags:
      - waste-controller
      summary: Selejtezés
      operationId: getCurrentTrash
      responses:
        "200":
          description: Sikeres lekérdezés
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Waste'
  /delete:
    delete:
      tags:
      - waste-controller
      summary: Kivételezés
      operationId: emptyTrash
      responses:
        "403":
          description: Nincs jogosultság
          content:
            application/json: {}
        "200":
          description: Sikeres törlés
          content:
            application/json: {}
        "401":
          description: Nincs bejelentkezve
          content:
            application/json: {}
        "405":
          description: Method not allowed
          content:
            application/json: {}
        "500":
          description: Nincs ilyen id-j? termék
          content:
            application/json: {}
      security:
      - apikey:
        - storage
      - openid:
        - storage
      - oauth2:
        - storage
components:
  schemas:
    Waste:
      type: object
      properties:
        id:
          type: integer
          format: int64
        removed:
          type: boolean
        name:
          type: string
        arrivedAt:
          type: string
          format: date-time
        price:
          type: number
          format: double
        count:
          type: integer
          format: int32
  securitySchemes:
    apikey:
      type: apiKey
      description: Keycloak Yokudlela
      name: Authorization
      in: header
    openid:
      type: openIdConnect
      description: Keycloak yokudlela
      openIdConnectUrl: http://yokudlela:6080/auth/realms/yokudlela/.well-known/openid-configuration
    oauth2:
      type: oauth2
      description: Keycloak Yokudlela
      flows:
        implicit:
          authorizationUrl: http://yokudlela:6080/auth/realms/yokudlela/protocl/openid-connect/auth?client_id=account&redirect_uri=http://localhost:8080/storage/swagger-ui/oauth2-redirect.html&response_type=code&scope=openid
          scopes: {}